Add files via upload
1. Importing the Tkinter module. 
2. Creation of the main window (container). 
3. Addition of widgets to the main window 
4. Applying the event Trigger on widgets like buttons, etc. 
     
Creating the File Open:
In order to do so, we have to import the filedialog module from Tkinter. The File dialog module will help you open, save files or directories.
In order to open a file explorer, we have to use the method, askopenfilename(). This function creates a file dialog object.